requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph == 0.17
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.8.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.7
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.1
executing==2.1.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
folium==0.17.0
fonttools==4.54.1
frozenlist==1.4.0
future == 0.18.2
geographiclib==2.0
geopy==2.4.1
ggshield==1.32.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
macholib == 1.15
MarkupSafe==2.1.5
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
matplotlib==3.9.2
matplotlib-inline==0.1.7
msgspec==0.18.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.26.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==3.0.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.6.1
pygitguardian==1.17.0
Pygments==2.18.0
PyJWT==2.6.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==12.5.1
rsa==4.9
seaborn==0.13.2
six == 1.16.0
sniffio==1.3.1
soupsieve==2.6
SpoofMAC==2.1.1
SQLAlchemy==2.0.35
stack-data==0.6.3
tk==0.1.0
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.3
wireless==0.3.3
WTForms==3.1.2
xyzservices==2024.9.0
yarl==1.9.2
zipp==3.18.1
gunicorn==20.1.0